requirements.txt
xlrd ==1.2.0
allure-pytest==2.13.2
openpyxl==3.1.2
pyaml==21.10.1
PyMySQL==1.0.2